add fixed classes
