+++ src/main/java/com/example/Lion.java
@@ -1,32 +1,34 @@
 package com.example;
+
 import java.util.List;
 
 public class Lion {
+    String sex;
+    boolean hasMane;
+    Feline feline;
 
-    private boolean hasMane;
-    private Predator predator;
-
-    public Lion(String sex, Predator predator) throws Exception {
+    public Lion(String sex, Feline feline) throws Exception {
         if ("Самец".equals(sex)) {
             hasMane = true;
         } else if ("Самка".equals(sex)) {
             hasMane = false;
         } else {
-            throw new Exception("Используйте допустимые значения пола животного - самей или самка");
+            throw new Exception("Используйте допустимые значения пола животного - самец или самка");
         }
-        this.predator = predator;
+
+        this.sex = sex;
+        this.feline = feline;
     }
 
     public int getKittens() {
-        return 1;
+        return feline.getKittens();
     }
 
     public boolean doesHaveMane() {
-
         return hasMane;
     }
 
     public List<String> getFood() throws Exception {
-        return predator.eatMeat();
+        return feline.eatMeat();
     }
 }
